+++ src/components/collection/container.styl
@@ -1,4 +1,4 @@
-@import "../../components/constants.styl"
+@import "../constants.styl"
 
 .container
   display: flex
@@ -63,3 +63,4 @@
   display: flex
   > * + *
     margin-left: 10px
+
